Replace C-style casts

+++ src/kern/irq_mgr_flex.h
@@ -58,7 +58,7 @@ public:
     if (pos < 0)
       pos = _max_irq;
 
-    if (pos >= (int)_max_irq)
+    if (pos >= _max_irq)
       {
         // add as last
         auto &e = _chips[_used++];
